di for shared preference & change folder name
