+++ app/build.gradle
@@ -53,20 +53,20 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.2'
     implementation 'androidx.activity:activity-compose:1.7.2'
-    implementation platform('androidx.compose:compose-bom:2023.09.01')
-    implementation 'androidx.compose.ui:ui:1.5.1'
-    implementation 'androidx.compose.ui:ui-graphics:1.5.1'
-    implementation 'androidx.compose.ui:ui-tooling-preview:1.5.1'
+    implementation platform('androidx.compose:compose-bom:2023.09.02')
+    implementation 'androidx.compose.ui:ui:1.5.2'
+    implementation 'androidx.compose.ui:ui-graphics:1.5.2'
+    implementation 'androidx.compose.ui:ui-tooling-preview:1.5.2'
     implementation 'androidx.compose.material3:material3:1.1.2'
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.work:work-runtime-ktx:2.8.1'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-    androidTestImplementation platform('androidx.compose:compose-bom:2023.09.01')
+    androidTestImplementation platform('androidx.compose:compose-bom:2023.09.02')
     androidTestImplementation 'androidx.compose.ui:ui-test-junit4'
     debugImplementation 'androidx.compose.ui:ui-tooling'
-    debugImplementation 'androidx.compose.ui:ui-test-manifest:1.5.1'
+    debugImplementation 'androidx.compose.ui:ui-test-manifest:1.5.2'
 
     // beacon
     implementation 'org.altbeacon:android-beacon-library:2.19.6'
@@ -90,7 +90,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-runtime-compose:2.6.2'
 
     // more icons
-    implementation "androidx.compose.material:material-icons-extended:1.5.1"
+    implementation "androidx.compose.material:material-icons-extended:1.5.2"
 
     // maps
     implementation 'com.google.maps.android:maps-compose:2.15.0'
@@ -104,4 +104,7 @@ dependencies {
     // destinations
     implementation 'io.github.raamcosta.compose-destinations:core:1.9.53'
     ksp 'io.github.raamcosta.compose-destinations:ksp:1.9.53'
+
+    // datastore (similar to SharedPreferences)
+    implementation "androidx.datastore:datastore-preferences:1.0.0"
 }
