+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         tools:targetApi="34"
         android:theme="@style/Theme.ShuttleTracker">
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.ShuttleTracker"
             android:launchMode="singleTask">
