+++ app/src/main/java/edu/rpi/shuttletracker/presentation/announcements/AnnouncementsScreen.kt
@@ -1,4 +1,4 @@
-package edu.rpi.shuttletracker.ui.announcements
+package edu.rpi.shuttletracker.presentation.announcements
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -24,7 +24,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import edu.rpi.shuttletracker.data.models.Announcement
-import edu.rpi.shuttletracker.ui.errors.CheckResponseError
+import edu.rpi.shuttletracker.presentation.errors.CheckResponseError
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Destination
@@ -59,7 +59,8 @@ fun AnnouncementsScreen(
         },
     ) { padding ->
         LazyColumn(
-            modifier = Modifier.padding(padding),
+            modifier = Modifier.padding(padding)
+                .padding(horizontal = 10.dp),
             verticalArrangement = Arrangement.spacedBy(15.dp),
         ) {
             itemsIndexed(announcementsUIState.announcements) { _, item ->
