+++ app/src/main/java/edu/rpi/shuttletracker/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package edu.rpi.shuttletracker.ui.theme
+package edu.rpi.shuttletracker.presentation.theme
 
 import androidx.compose.ui.graphics.Color
