+++ app/src/main/java/edu/rpi/shuttletracker/presentation/theme/Type.kt
@@ -1,4 +1,4 @@
-package edu.rpi.shuttletracker.ui.theme
+package edu.rpi.shuttletracker.presentation.theme
 
 import androidx.compose.material3.Typography
